requirements.txt
mysql-connector-python==8.0.33
tabulate==0.9.0
python-dotenv==1.0.0